Remove '' category and Invalid version warnings
library.properties currently does not include the category keyword. This causes Arduino IDE to create a warning: 
WARNING: Category '' in library UIPEthernet is not valid. Setting to 'Uncategorized'

Version string 1.04 also causes a warning. This is a major nuisance because the warning message keeps repeating itself:
Invalid version found: 1.04

Arduino IDE expects the version number to be in the 1.0.4 format instead of 1.04. 1.04 causes library to report "invalid" version. Hence the warning.